[IMP] Auto-Update working

+++ update_server/wizard/manage_version.py
@@ -8,24 +8,26 @@ Created on 9 juil. 2012
 from osv import osv
 from osv import fields
 import tools
-import pprint
-import sync_server
-pp = pprint.PrettyPrinter(indent=4)
-import logging
+#import pprint
+#import sync_server
+#pp = pprint.PrettyPrinter(indent=4)
+#import logging
 from datetime import datetime
 from dateutil.relativedelta import relativedelta
+import hashlib
+from base64 import b64decode
 
 class manage_version(osv.osv):
     
     _name = "sync_server.version.manager"
     
     _columns = {
-        'name' : fields.char('Revision Tag', size=256),
-        'hash' : fields.char('Commit Hash', size=512),
-        'date' : fields.datetime('Revision Date', readonly=True),
-        'version_ids' : fields.many2many('sync_server.version', 'sync_server_version_rel', 'wiz_id', 'version_id', string="History of revision", readonly=True, limit=10000),
-        'create_date' : fields.datetime("Create date"),
-               
+        'name' : fields.char('Revision', size=256),
+        'patch' : fields.binary('Patch'),
+        'date' : fields.datetime('Date', readonly=True),
+        'version_ids' : fields.many2many('sync_server.version', 'sync_server_version_rel', 'wiz_id', 'version_id', string="History of Revision", readonly=True, limit=10000),
+        'create_date' : fields.datetime("Create Date"),
+        'importance' : fields.selection([('required','Required'),('optional','Optional')], "Importance Flag"),
     }
     
     def _get_version(self, cr, uid, context=None):
@@ -38,25 +40,31 @@ class manage_version(osv.osv):
     
     def add_revision(self, cr, uid, ids, context=None):
         for wiz in self.browse(cr, uid, ids, context=context):
+            m = hashlib.md5()
+            m.update( b64decode(wiz.patch) )
             data = {
                 'name' :  wiz.name,
-                'commit' : wiz.hash,
+                'sum' : m.hexdigest(),
                 'date' : wiz.date,
+                'patch' : wiz.patch,
+                'importance' : wiz.importance,
             }
             res_id = self.pool.get("sync_server.version").create(cr, uid, data, context=context)
             self.write(cr, uid, [wiz.id], {'version_ids' : [(4, res_id)], 
                                            'name' : False, 
-                                           'hash' : False, 
+                                           'patch' : False,
+                                           'importance' : False,
                                            'date' : datetime.now().strftime("%Y-%m-%d %H:%M:%S")}, 
                        context=context)
         return True
             
             
     def vacuum(self, cr, uid):
-        today = (datetime.now() + relativedelta(hours=-1)).strftime("%Y-%m-%d %H:%M:%S") 
-        unlink_ids = self.search(cr, uid, [('create_date', '<', today)])
-        print "auto Vacuum", unlink_ids
-        self.unlink(cr, uid, unlink_ids)
+        now = (datetime.now() + relativedelta(hours=-1)).strftime("%Y-%m-%d %H:%M:%S") 
+        unlink_ids = self.search(cr, uid, [('create_date', '<', now)])
+        if unlink_ids:
+            print "auto Vacuum", unlink_ids
+            self.unlink(cr, uid, unlink_ids)
         return True
